qemugen: Fix handling of discard and detect-zeroes options for raw disk files
While the generated file blockdev was properly configured to unmap blocks,
the raw blockdev layer that sits on top of it was not. This effectively
blocked all TRIM/discard/unmap commands coming from the guest, preventing
them from passing through to the underlying storage device.

This commit configures the raw blockdev properly, causing it to pass
the TRIM/discard/unmap commands down to the file blockdev.

Additionally, this commit moves the detect-zeroes option from the file
blockdev to the raw blockdev. Presumably, only one layer in the block device
node graph needs to perform the detection. It's plausible that having it
enabled in both blockdev nodes would cause the detection to happen twice,
which would have a negative impact on performance.